test : 노드 시스템 UI 추가 및 PlayerStat에 데이터 전송

+++ ProjectP/Assets/02.Scripts/StatNode/Node/NodeInfo.cs
@@ -8,7 +8,7 @@ public class NodeInfo: ISheetParsable, IIdentifiable
     private int _id;
     private int _nodeLevel;
     private int _nodeCostPoint;
-    private int _nodeIncrValue;
+    private float _nodeIncrValue;
     private string _description;
     
     // getter
@@ -17,7 +17,7 @@ public class NodeInfo: ISheetParsable, IIdentifiable
     public int Id => _id;
     public int NodeLevel => _nodeLevel;
     public int NodeCostPoint => _nodeCostPoint;
-    public int NodeIncrValue => _nodeIncrValue;
+    public float NodeIncrValue => _nodeIncrValue;
     public string Description => _description;
     
     [field: SerializeField] public string Name { get; set; }
@@ -35,7 +35,7 @@ public class NodeInfo: ISheetParsable, IIdentifiable
         int.TryParse(data[2], out _id);    // C열 (ID)
         int.TryParse(data[3], out _nodeLevel); // D열 노드 레벨
         int.TryParse(data[4], out _nodeCostPoint); // E열 노드 코스트 포인트
-        int.TryParse(data[5], out _nodeIncrValue); // F열 증감치
+        float.TryParse(data[5], out _nodeIncrValue); // F열 증감치
         _description = data[6];            // G열 설명
     }
 }
